add socket io init
